+++ pytsite/content/_functions.py
@@ -4,28 +4,31 @@ __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-from datetime import datetime
+from datetime import datetime as _datetime
 from pytsite import admin as _admin, taxonomy as _taxonomy
 from pytsite.core import odm as _odm, util as _util, router as _router, lang as _lang
 from . import _model
 
-__models = []
+__models = {}
 
 
-def register_model(model: str, cls, menu_title: str, menu_weight: int=0, menu_icon: str='fa fa-file-text-o'):
+def register_model(model: str, cls, title: str, menu_weight: int=0, menu_icon: str='fa fa-file-text-o'):
     """Register content model.
     """
     if isinstance(cls, str):
         cls = _util.get_class(cls)
 
     if not issubclass(cls, _model.Content):
-        raise TypeError('Subclass of ContentModel expected.')
+        raise TypeError('Subclass of content model expected.')
+
+    if is_model_registered(model):
+        raise KeyError("Model '{}' is already registered.".format(model))
 
     _odm.register_model(model, cls)
-    __models.append(model)
+    __models[model] = (cls, title)
 
     menu_url = _router.endpoint_url('pytsite.odm_ui.eps.browse', {'model': model})
-    _admin.sidebar.add_menu('content', model, menu_title, menu_url, menu_icon, weight=menu_weight, permissions=(
+    _admin.sidebar.add_menu('content', model, title, menu_url, menu_icon, weight=menu_weight, permissions=(
         'pytsite.odm_ui.browse.' + model,
         'pytsite.odm_ui.browse_own.' + model,
     ))
@@ -37,29 +40,38 @@ def is_model_registered(model: str) -> bool:
     return model in __models
 
 
-def get_registered_models() -> list:
+def get_models() -> dict:
     """Get registered content models.
     """
     return __models
 
 
-def find(model: str, status='published', check_publish_time=True):
-    """Get entity finder.
+def get_model(model: str) -> tuple:
+    """Get model information.
     """
     if not is_model_registered(model):
-        raise Exception("Model '{}' is not registered as content model.".format(model))
+        raise KeyError("Model '{}' is not registered as content model.".format(model))
+
+    return __models[model]
+
+
+def find(model: str, status='published', check_publish_time=True):
+    """Get content entities finder.
+    """
+    if not is_model_registered(model):
+        raise KeyError("Model '{}' is not registered as content model.".format(model))
 
     f = _odm.find(model).sort([('publish_time', _odm.I_DESC)])
     if status:
         f.where('status', '=', status)
     if check_publish_time:
-        f.where('publish_time', '<=', datetime.now())
+        f.where('publish_time', '<=', _datetime.now())
 
     return f
 
 
 def get_publish_statuses() -> list:
-    """Get content publish statuses.
+    """Get allowed content publication statuses.
     """
     r = []
     for s in ('published', 'waiting', 'unpublished'):