+++ pytsite/taxonomy/_widget.py
@@ -86,7 +86,7 @@ class Cloud(_widget.Base):
         """
         root = _html.Div(child_separator='  ')
         weight = 10
-        for term in _functions.find(self._model).get(self._num):
+        for term in self._get_finder().get(self._num):
             title = self._title_pattern % term.f_get('title')
             a_cls = 'term {} weight-{} {}'.format(self._model, weight, self._term_css)
             a = _html.A(title, href=self._link_pattern % term.f_get('alias'), cls=a_cls)
@@ -94,3 +94,6 @@ class Cloud(_widget.Base):
             weight -= 1
 
         return self._group_wrap(root)
+
+    def _get_finder(self):
+        _functions.find(self._model)